fix: Change physics constants to same type as in VPX.

+++ VisualPinball.Unity/VisualPinball.Unity/VPT/Ball/BallVelocitySystem.cs
@@ -47,7 +47,7 @@ namespace VisualPinball.Unity
 
 				marker.Begin();
 
-				ball.Velocity += gravity * PhysicsConstants.PhysFactor;
+				ball.Velocity += gravity * (float)PhysicsConstants.PhysFactor;
 
 				marker.End();
